+++ entity/src/epilot/models/operations/createentity.py
@@ -11,7 +11,7 @@ from typing import Optional
 @dataclasses.dataclass
 class CreateEntityRequest:
     slug: str = dataclasses.field(metadata={'path_param': { 'field_name': 'slug', 'style': 'simple', 'explode': False }})
-    r"""Entity Schema"""
+    r"""Entity Type"""
     activity_id: Optional[str] = dataclasses.field(default=None, metadata={'query_param': { 'field_name': 'activity_id', 'style': 'form', 'explode': True }})
     r"""Activity to include in event feed"""
     async_: Optional[bool] = dataclasses.field(default=False, metadata={'query_param': { 'field_name': 'async', 'style': 'form', 'explode': True }})