+++ entity/src/epilot/models/operations/getschema.py
@@ -10,6 +10,7 @@ from typing import Optional
 @dataclasses.dataclass
 class GetSchemaRequest:
     slug: str = dataclasses.field(metadata={'path_param': { 'field_name': 'slug', 'style': 'simple', 'explode': False }})
+    r"""Entity Type"""
     id: Optional[str] = dataclasses.field(default=None, metadata={'query_param': { 'field_name': 'id', 'style': 'form', 'explode': True }})