+++ entity/src/epilot/models/operations/getschemaversions.py
@@ -12,6 +12,7 @@ from typing import List, Optional
 @dataclasses.dataclass
 class GetSchemaVersionsRequest:
     slug: str = dataclasses.field(metadata={'path_param': { 'field_name': 'slug', 'style': 'simple', 'explode': False }})
+    r"""Entity Type"""
     drafts_from: Optional[float] = dataclasses.field(default=None, metadata={'query_param': { 'field_name': 'drafts_from', 'style': 'form', 'explode': True }})
     drafts_size: Optional[float] = dataclasses.field(default=None, metadata={'query_param': { 'field_name': 'drafts_size', 'style': 'form', 'explode': True }})
     versions_from: Optional[float] = dataclasses.field(default=None, metadata={'query_param': { 'field_name': 'versions_from', 'style': 'form', 'explode': True }})