+++ entity/src/epilot/models/operations/gettaxonomy.py
@@ -10,7 +10,7 @@ from typing import Optional
 @dataclasses.dataclass
 class GetTaxonomyRequest:
     taxonomy_slug: str = dataclasses.field(metadata={'path_param': { 'field_name': 'taxonomySlug', 'style': 'simple', 'explode': False }})
-    r"""Taxonomy slug to return taxonomy for"""
+    r"""Taxonomy slug"""