+++ entity/src/epilot/models/shared/entitysearchparams.py
@@ -21,7 +21,7 @@ class EntitySearchParams:
     r"""Lucene queries supported with ElasticSearch"""
     aggs: Optional[EntitySearchParamsAggs] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('aggs'), 'exclude': lambda f: f is None }})
     r"""Aggregation supported by ElasticSearch allows summarizing data as metrics, statistics, or other analytics."""
-    fields_: Optional[List[str]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('fields'), 'exclude': lambda f: f is None }})
+    fields: Optional[List[str]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('fields'), 'exclude': lambda f: f is None }})
     r"""List of entity fields to include in search results"""
     from_: Optional[int] = dataclasses.field(default=0, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('from'), 'exclude': lambda f: f is None }})
     hydrate: Optional[bool] = dataclasses.field(default=False, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('hydrate'), 'exclude': lambda f: f is None }})